email_service/requirements.txt
Flask
mailjet_rest
load_dotenv
pika
APScheduler
requests